--- schtasks/examples/three_times_a_day.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-16"?>
-<Task version="1.2" xmlns="http://schemas.microsoft.com/windows/2004/02/mit/task">
-  <RegistrationInfo>
-    <Date>2025-02-06T22:46:29</Date>
-    <Author>resticprofile</Author>
-    <Description>three times a day</Description>
-    <URI>\resticprofile backup\test 6</URI>
-  </RegistrationInfo>
-  <Principals>
-    <Principal id="Author">
-      <UserId>S-1-5-21-3622340250-3004114779-3108145013-1002</UserId>
-      <LogonType>InteractiveToken</LogonType>
-    </Principal>
-  </Principals>
-  <Settings>
-    <DisallowStartIfOnBatteries>true</DisallowStartIfOnBatteries>
-    <StopIfGoingOnBatteries>true</StopIfGoingOnBatteries>
-    <MultipleInstancesPolicy>IgnoreNew</MultipleInstancesPolicy>
-    <IdleSettings>
-      <Duration>PT10M</Duration>
-      <WaitTimeout>PT1H</WaitTimeout>
-      <StopOnIdleEnd>true</StopOnIdleEnd>
-      <RestartOnIdle>false</RestartOnIdle>
-    </IdleSettings>
-  </Settings>
-  <Triggers>
-    <CalendarTrigger>
-      <StartBoundary>2025-02-07T03:04:00</StartBoundary>
-      <ExecutionTimeLimit>PT0S</ExecutionTimeLimit>
-      <Repetition>
-        <Interval>PT1H</Interval>
-        <Duration>PT2H</Duration>
-      </Repetition>
-      <ScheduleByDay>
-        <DaysInterval>1</DaysInterval>
-      </ScheduleByDay>
-    </CalendarTrigger>
-  </Triggers>
-  <Actions Context="Author">
-    <Exec>
-      <Command>echo</Command>
-      <Arguments>hello there</Arguments>
-      <WorkingDirectory>C:\</WorkingDirectory>
-    </Exec>
-  </Actions>
-</Task>
