+++ schtasks/settings.go
@@ -1,27 +1,27 @@
 package schtasks
 
-import "github.com/rickb777/date/period"
+import "github.com/rickb777/period"
 
 // Settings provides the settings that the Task Scheduler service uses to perform the task
 // https://docs.microsoft.com/en-us/windows/desktop/api/taskschd/nn-taskschd-itasksettings
 type Settings struct {
-	AllowStartOnDemand         bool              `xml:"AllowStartOnDemand"` // indicates that the task can be started by using either the Run command or the Context menu
-	AllowHardTerminate         bool              `xml:"AllowHardTerminate"` // indicates that the task may be terminated by the Task Scheduler service using TerminateProcess
+	AllowStartOnDemand         bool              `xml:"AllowStartOnDemand,omitempty"` // indicates that the task can be started by using either the Run command or the Context menu
+	AllowHardTerminate         bool              `xml:"AllowHardTerminate,omitempty"` // indicates that the task may be terminated by the Task Scheduler service using TerminateProcess
 	Compatibility              Compatibility     // indicates which version of Task Scheduler a task is compatible with
 	DeleteExpiredTaskAfter     *period.Period    `xml:"DeleteExpiredTaskAfter,omitempty"` // the amount of time that the Task Scheduler will wait before deleting the task after it expires
 	DisallowStartIfOnBatteries bool              `xml:"DisallowStartIfOnBatteries"`       // indicates that the task will not be started if the computer is running on batteries
 	ExecutionTimeLimit         period.Period     `xml:"ExecutionTimeLimit"`               // the amount of time that is allowed to complete the task
-	Hidden                     bool              `xml:"Hidden"`                           // indicates that the task will not be visible in the UI
+	Hidden                     bool              `xml:"Hidden,omitempty"`                 // indicates that the task will not be visible in the UI
 	IdleSettings               IdleSettings      `xml:"IdleSettings"`
 	MultipleInstancesPolicy    InstancesPolicy   `xml:"MultipleInstancesPolicy"` // defines how the Task Scheduler deals with multiple instances of the task
-	Priority                   uint              `xml:"Priority"`                // the priority level of the task, ranging from 0 - 10, where 0 is the highest priority, and 10 is the lowest. Only applies to ComHandler, Email, and MessageBox actions
+	Priority                   uint              `xml:"Priority,omitempty"`      // the priority level of the task, ranging from 0 - 10, where 0 is the highest priority, and 10 is the lowest. Only applies to ComHandler, Email, and MessageBox actions
 	RestartOnFailure           *RestartOnFailure `xml:"RestartOnFailure,omitempty"`
-	RunOnlyIfIdle              bool              `xml:"RunOnlyIfIdle"`             // indicates that the Task Scheduler will run the task only if the computer is in an idle condition
-	RunOnlyIfNetworkAvailable  bool              `xml:"RunOnlyIfNetworkAvailable"` // indicates that the Task Scheduler will run the task only when a network is available
-	StartWhenAvailable         bool              `xml:"StartWhenAvailable"`        // indicates that the Task Scheduler can start the task at any time after its scheduled time has passed
-	StopIfGoingOnBatteries     bool              `xml:"StopIfGoingOnBatteries"`    // indicates that the task will be stopped if the computer is going onto batteries
-	WakeToRun                  bool              `xml:"WakeToRun"`                 // indicates that the Task Scheduler will wake the computer when it is time to run the task, and keep the computer awake until the task is completed
-	UseUnifiedSchedulingEngine bool              `xml:"UseUnifiedSchedulingEngine"`
+	RunOnlyIfIdle              bool              `xml:"RunOnlyIfIdle,omitempty"`             // indicates that the Task Scheduler will run the task only if the computer is in an idle condition
+	RunOnlyIfNetworkAvailable  bool              `xml:"RunOnlyIfNetworkAvailable,omitempty"` // indicates that the Task Scheduler will run the task only when a network is available
+	StartWhenAvailable         bool              `xml:"StartWhenAvailable,omitempty"`        // indicates that the Task Scheduler can start the task at any time after its scheduled time has passed
+	StopIfGoingOnBatteries     bool              `xml:"StopIfGoingOnBatteries,omitempty"`    // indicates that the task will be stopped if the computer is going onto batteries
+	WakeToRun                  bool              `xml:"WakeToRun,omitempty"`                 // indicates that the Task Scheduler will wake the computer when it is time to run the task, and keep the computer awake until the task is completed
+	UseUnifiedSchedulingEngine bool              `xml:"UseUnifiedSchedulingEngine,omitempty"`
 }
 
 // IdleSettings specifies how the Task Scheduler performs tasks when the computer is in an idle condition.
