+++ schtasks/task_test.go
@@ -1,90 +1,46 @@
 package schtasks
 
 import (
-	"encoding/xml"
-	"io"
-	"os"
-	"path/filepath"
 	"testing"
+	"time"
 
-	"github.com/creativeprojects/resticprofile/constants"
+	"github.com/creativeprojects/resticprofile/calendar"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func TestLoadXMLTask(t *testing.T) {
-	filenames, err := filepath.Glob("examples/*.xml")
-	require.NoError(t, err)
-	for _, filename := range filenames {
-		file, err := os.Open(filename)
+func TestCompileDifferences(t *testing.T) {
+	testData := []struct {
+		input       string
+		differences []time.Duration
+		unique      []time.Duration
+	}{
+		{
+			"1..4,6,8:00",
+			[]time.Duration{1 * time.Hour, 1 * time.Hour, 1 * time.Hour, 2 * time.Hour, 2 * time.Hour},
+			[]time.Duration{1 * time.Hour, 2 * time.Hour},
+		},
+		{
+			"Sat,Sun 0,12:00",
+			[]time.Duration{12 * time.Hour},
+			[]time.Duration{12 * time.Hour},
+		},
+		{
+			"mon *-11..12-* 1,13:00",
+			[]time.Duration{12 * time.Hour},
+			[]time.Duration{12 * time.Hour},
+		},
+	}
+
+	for _, testItem := range testData {
+		event := calendar.NewEvent()
+		err := event.Parse(testItem.input)
 		require.NoError(t, err)
-		defer file.Close()
-
-		decoder := xml.NewDecoder(file)
-		decoder.CharsetReader = func(charset string, input io.Reader) (io.Reader, error) {
-			// no need for character conversion
-			return input, nil
-		}
-		task := Task{}
-		err = decoder.Decode(&task)
-		require.NoErrorf(t, err, "filename: %s", filename)
-		assert.True(t, len(task.Triggers.CalendarTrigger) > 0 || len(task.Triggers.TimeTrigger) > 0)
-
-		// t.Logf("%+v", task)
-		// t.Logf("%+v", task.Triggers.CalendarTrigger[0].ScheduleByWeek)
-		// t.Logf("%+v", task.Triggers.CalendarTrigger[0].ScheduleByMonth)
-		// t.Logf("%+v", task.Triggers.CalendarTrigger[0].ScheduleByMonthDayOfWeek)
+		ref, err := time.Parse(time.ANSIC, "Mon Jan 2 12:00:00 2006")
+		require.NoError(t, err)
+		start := event.Next(ref)
+		diff, uniques := compileDifferences(event.GetAllInBetween(start, start.Add(24*time.Hour)))
+		assert.ElementsMatch(t, testItem.differences, diff, "duration between triggers")
+		assert.ElementsMatch(t, testItem.unique, uniques, "unique set of durations between triggers")
 	}
-}
-
-func TestSaveXMLTask(t *testing.T) {
-	file, err := os.Create("output.xml")
-	require.NoError(t, err)
-	defer file.Close()
-
-	encoder := xml.NewEncoder(file)
-	encoder.Indent("", "  ")
-	task := NewTask()
-	task.Actions.Exec = []ExecAction{
-		{
-			Command:   "echo",
-			Arguments: "Hello World!",
-		},
-	}
-	task.Triggers.CalendarTrigger = []CalendarTrigger{
-		{
-			ScheduleByWeek: &ScheduleByWeek{
-				DaysOfWeek: DaysOfWeek{
-					Monday:    WeekDay,
-					Wednesday: WeekDay,
-				},
-				WeeksInterval: 1,
-			},
-			StartBoundary: "2020-01-02T03:04:00Z",
-		},
-	}
-	err = encoder.Encode(&task)
-	require.NoError(t, err)
-}
-
-func TestSaveXMLTaskUsingServiceAccount(t *testing.T) {
-	file, err := os.Create("output.xml")
-	require.NoError(t, err)
-	defer file.Close()
-
-	encoder := xml.NewEncoder(file)
-	encoder.Indent("", "  ")
-	task := NewTask()
-	task.RegistrationInfo.Author = constants.ApplicationName
-	task.Principals.Principal.UserId = serviceAccount
-	task.Principals.Principal.RunLevel = RunLevelLeastPrivilege
-	task.Principals.Principal.LogonType = ""
-	task.Actions.Exec = []ExecAction{
-		{
-			Command:   "echo",
-			Arguments: "Hello World!",
-		},
-	}
-	err = encoder.Encode(&task)
-	require.NoError(t, err)
 }
