+++ schtasks/taskscheduler.go
@@ -1,627 +1,217 @@
-//go:build windows && taskmaster
-
-package schtasks
-
-import (
-	"errors"
-	"fmt"
-	"strings"
-	"text/tabwriter"
-	"time"
-
-	"github.com/capnspacehook/taskmaster"
-	"github.com/creativeprojects/clog"
-	"github.com/creativeprojects/resticprofile/calendar"
-	"github.com/creativeprojects/resticprofile/constants"
-	"github.com/creativeprojects/resticprofile/term"
-	"github.com/rickb777/date/period"
-)
-
 // Schedule types on Windows:
 // ==========================
 // 1. one time:
-//    - at a specific date
+//   - at a specific date
+//
 // 2. daily:
-//    - 1 start date
-//    - recurring every n days
+//   - 1 start date
+//   - recurring every n days
+//
 // 3. weekly:
-//    - 1 start date
-//    - recurring every n weeks
-//    - on specific weekdays
+//   - 1 start date
+//   - recurring every n weeks
+//   - on specific weekdays
+//
 // 4. monthly:
-//    - 1 start date
-//    - on specific months
-//    - on specific days (1 to 31)
+//   - 1 start date
+//   - on specific months
+//   - on specific days (1 to 31)
+package schtasks
+
+import (
+	"bytes"
+	"fmt"
+	"os"
+	"os/exec"
+	"slices"
+	"strings"
+	"text/tabwriter"
+
+	"github.com/creativeprojects/clog"
+	"github.com/creativeprojects/resticprofile/calendar"
+	"github.com/creativeprojects/resticprofile/term"
+)
 
 const (
-	binaryPath     = "schtasks.exe"
-	systemUserName = "SYSTEM"
+	tasksPath  = `\resticprofile backup\`
+	binaryPath = "schtasks.exe"
+	// From: https://learn.microsoft.com/en-us/windows/win32/secauthz/security-descriptor-string-format
+	// O:owner_sid
+	// G:group_sid
+	// D:dacl_flags(string_ace1)(string_ace2)... (string_acen)  <---
+	// S:sacl_flags(string_ace1)(string_ace2)... (string_acen)
+	// With flag:
+	// "AI"	SDDL_AUTO_INHERITED
+	// From: https://learn.microsoft.com/en-us/windows/win32/secauthz/ace-strings
+	// - first field:
+	// "A" 	SDDL_ACCESS_ALLOWED
+	// - third field
+	// "FA" 	SDDL_FILE_ALL 	FILE_GENERIC_ALL
+	// "FR" 	SDDL_FILE_READ 	FILE_GENERIC_READ
+	// "FW" 	SDDL_FILE_WRITE 	FILE_GENERIC_WRITE
+	// "FX" 	SDDL_FILE_EXECUTE 	FILE_GENERIC_EXECUTE
+	// From: https://learn.microsoft.com/en-us/windows/win32/secauthz/sid-strings
+	// "AU" 	SDDL_AUTHENTICATED_USERS
+	// "BA" 	SDDL_BUILTIN_ADMINISTRATORS
+	// "LS" 	SDDL_LOCAL_SERVICE
+	// "SY" 	SDDL_LOCAL_SYSTEM
+	securityDescriptor = "D:AI(A;;FA;;;BA)(A;;FA;;;SY)(A;;FRFX;;;LS)(A;;FR;;;AU)"
 )
 
-var (
-	// no need to recreate the service every time
-	taskService taskmaster.TaskService
-)
-
-// ErrNotConnected is returned by public functions if Connect was not called, was not successful or Close closed the connection.
-var ErrNotConnected = errors.New("local task scheduler not connected")
-
-// IsConnected returns whether a connection to the local task scheduler is established
-func IsConnected() bool {
-	return taskService.IsConnected()
-}
-
-// Connect initializes a connection to the local task scheduler
-func Connect() error {
-	var err error
-
-	if !IsConnected() {
-		taskService, err = taskmaster.Connect()
+// Init only checks the schtask.exe tool is available
+func Init() error {
+	found, err := exec.LookPath(binaryPath)
+	if err != nil || found == "" {
+		return fmt.Errorf("it doesn't look like %s is installed on your system", binaryPath)
 	}
-	return err
-}
-
-// Close releases the resources used by the task service
-func Close() {
-	taskService.Disconnect()
+	return nil
 }
 
 // Create or update a task (if the name already exists in the Task Scheduler)
 func Create(config *Config, schedules []*calendar.Event, permission Permission) error {
-	if !IsConnected() {
-		return ErrNotConnected
+	var (
+		// username, password string
+		err error
+	)
+	task := createTaskDefinition(config, schedules)
+	taskPath := getTaskPath(config.ProfileName, config.CommandName)
+	task.RegistrationInfo.URI = taskPath
+
+	switch permission {
+	case SystemAccount:
+		task.Principals.Principal.LogonType = LogonTypeServiceForUser
+		task.Principals.Principal.RunLevel = RunLevelHighest
+		task.Principals.Principal.UserId = serviceAccount
+		task.RegistrationInfo.SecurityDescriptor = securityDescriptor // allow authenticated users to read the task status
+
+	case UserLoggedOnAccount:
+		task.Principals.Principal.LogonType = LogonTypeInteractiveToken
+
+	default:
+		task.Principals.Principal.LogonType = LogonTypePassword
+		// username, password, err = userCredentials()
+		// if err != nil {
+		// 	return fmt.Errorf("cannot get user name or password: %w", err)
+		// }
 	}
 
-	if permission == SystemAccount {
-		return createSystemTask(config, schedules)
-	}
-	if permission == UserLoggedOnAccount {
-		return createUserLoggedOnTask(config, schedules)
-	}
-	return createUserTask(config, schedules)
-}
-
-// createUserTask creates a new user task. Will update an existing task instead of overwritting
-func createUserTask(config *Config, schedules []*calendar.Event) error {
-	taskName := getTaskPath(config.ProfileName, config.CommandName)
-	registeredTask, err := taskService.GetRegisteredTask(taskName)
-	if err == nil {
-		// the task already exists
-		return updateUserTask(registeredTask, config, schedules)
-	}
-
-	username, password, err := userCredentials()
+	file, err := os.CreateTemp("", "*.xml")
 	if err != nil {
-		return fmt.Errorf("cannot get user name or password: %w", err)
+		return fmt.Errorf("cannot create XML task file: %w", err)
 	}
+	filename := file.Name()
+	defer func() {
+		file.Close()
+		os.Remove(filename)
+	}()
 
-	task := taskService.NewTaskDefinition()
-	task.AddAction(taskmaster.ExecAction{
-		Path:       config.Command,
-		Args:       config.Arguments,
-		WorkingDir: config.WorkingDirectory,
-	})
-	task.Principal.LogonType = taskmaster.TASK_LOGON_PASSWORD
-	task.Principal.RunLevel = taskmaster.TASK_RUNLEVEL_LUA
-	task.Principal.UserID = username
-	task.RegistrationInfo.Author = constants.ApplicationName
-	task.RegistrationInfo.Description = config.JobDescription
-
-	createSchedules(&task, schedules)
-
-	_, created, err := taskService.CreateTaskEx(
-		taskName,
-		task,
-		username,
-		password,
-		taskmaster.TASK_LOGON_PASSWORD,
-		false)
+	err = createTaskFile(task, file)
 	if err != nil {
-		return err
+		return fmt.Errorf("cannot write XML task file: %w", err)
 	}
-	if !created {
-		return errors.New("cannot create user task")
+	file.Close()
+
+	stdout, stderr := &bytes.Buffer{}, &bytes.Buffer{}
+	cmd := exec.Command(binaryPath, "/create", "/tn", taskPath, "/xml", filename)
+	cmd.Stdout = stdout
+	cmd.Stderr = stderr
+	err = cmd.Run()
+	if err != nil {
+		return schTasksError(stderr.String())
 	}
 	return nil
-}
-
-// updateUserTask updates an existing task
-func updateUserTask(task taskmaster.RegisteredTask, config *Config, schedules []*calendar.Event) error {
-	taskName := getTaskPath(config.ProfileName, config.CommandName)
-
-	username, password, err := userCredentials()
-	if err != nil {
-		return fmt.Errorf("cannot get user name or password: %w", err)
-	}
-
-	// clear up all actions and put ours back
-	task.Definition.Actions = make([]taskmaster.Action, 0, 1)
-	task.Definition.AddAction(taskmaster.ExecAction{
-		Path:       config.Command,
-		Args:       config.Arguments,
-		WorkingDir: config.WorkingDirectory,
-	})
-	task.Definition.Principal.LogonType = taskmaster.TASK_LOGON_PASSWORD
-	task.Definition.Principal.RunLevel = taskmaster.TASK_RUNLEVEL_LUA
-	task.Definition.Principal.UserID = username
-
-	// clear up all schedules and put them back
-	task.Definition.Triggers = []taskmaster.Trigger{}
-	createSchedules(&task.Definition, schedules)
-
-	_, err = taskService.UpdateTaskEx(
-		taskName,
-		task.Definition,
-		username,
-		password,
-		taskmaster.TASK_LOGON_PASSWORD)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// createUserLoggedOnTask creates a new user task. Will update an existing task instead of overwriting
-func createUserLoggedOnTask(config *Config, schedules []*calendar.Event) error {
-	taskName := getTaskPath(config.ProfileName, config.CommandName)
-	registeredTask, err := taskService.GetRegisteredTask(taskName)
-	if err == nil {
-		// the task already exists
-		return updateUserLoggedOnTask(registeredTask, config, schedules)
-	}
-
-	task := taskService.NewTaskDefinition()
-	task.AddAction(taskmaster.ExecAction{
-		Path:       config.Command,
-		Args:       config.Arguments,
-		WorkingDir: config.WorkingDirectory,
-	})
-	task.Principal.LogonType = taskmaster.TASK_LOGON_INTERACTIVE_TOKEN
-	task.Principal.RunLevel = taskmaster.TASK_RUNLEVEL_LUA
-	task.RegistrationInfo.Author = constants.ApplicationName
-	task.RegistrationInfo.Description = config.JobDescription
-
-	createSchedules(&task, schedules)
-
-	_, created, err := taskService.CreateTask(taskName, task, false)
-	if err != nil {
-		return fmt.Errorf("cannot create user task: %w", err)
-	}
-	if !created {
-		return errors.New("cannot create user task")
-	}
-	return nil
-}
-
-// updateUserLoggedOnTask updates an existing task
-func updateUserLoggedOnTask(task taskmaster.RegisteredTask, config *Config, schedules []*calendar.Event) error {
-	taskName := getTaskPath(config.ProfileName, config.CommandName)
-
-	// clear up all actions and put ours back
-	task.Definition.Actions = make([]taskmaster.Action, 0, 1)
-	task.Definition.AddAction(taskmaster.ExecAction{
-		Path:       config.Command,
-		Args:       config.Arguments,
-		WorkingDir: config.WorkingDirectory,
-	})
-	task.Definition.Principal.LogonType = taskmaster.TASK_LOGON_INTERACTIVE_TOKEN
-	task.Definition.Principal.RunLevel = taskmaster.TASK_RUNLEVEL_LUA
-
-	// clear up all schedules and put them back
-	task.Definition.Triggers = []taskmaster.Trigger{}
-	createSchedules(&task.Definition, schedules)
-
-	_, err := taskService.UpdateTaskEx(
-		taskName,
-		task.Definition,
-		"",
-		"",
-		taskmaster.TASK_LOGON_INTERACTIVE_TOKEN)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// createSystemTask creates a new system task. Will update an existing task instead of overwritting
-func createSystemTask(config *Config, schedules []*calendar.Event) error {
-	taskName := getTaskPath(config.ProfileName, config.CommandName)
-	registeredTask, err := taskService.GetRegisteredTask(taskName)
-	if err == nil {
-		// the task already exists
-		return updateSystemTask(registeredTask, config, schedules)
-	}
-
-	task := taskService.NewTaskDefinition()
-	task.AddAction(taskmaster.ExecAction{
-		Path:       config.Command,
-		Args:       config.Arguments,
-		WorkingDir: config.WorkingDirectory,
-	})
-	task.Principal.LogonType = taskmaster.TASK_LOGON_SERVICE_ACCOUNT
-	task.Principal.RunLevel = taskmaster.TASK_RUNLEVEL_HIGHEST
-	task.Principal.UserID = systemUserName
-	task.RegistrationInfo.Author = constants.ApplicationName
-	task.RegistrationInfo.Description = config.JobDescription
-
-	createSchedules(&task, schedules)
-
-	_, created, err := taskService.CreateTask(taskName, task, false)
-	if err != nil {
-		return err
-	}
-	if !created {
-		return errors.New("cannot create system task")
-	}
-	return nil
-}
-
-// updateSystemTask updates an existing task
-func updateSystemTask(task taskmaster.RegisteredTask, config *Config, schedules []*calendar.Event) error {
-	taskName := getTaskPath(config.ProfileName, config.CommandName)
-
-	// clear up all actions and put ours back
-	task.Definition.Actions = make([]taskmaster.Action, 0, 1)
-	task.Definition.AddAction(taskmaster.ExecAction{
-		Path:       config.Command,
-		Args:       config.Arguments,
-		WorkingDir: config.WorkingDirectory,
-	})
-	task.Definition.Principal.LogonType = taskmaster.TASK_LOGON_SERVICE_ACCOUNT
-	task.Definition.Principal.RunLevel = taskmaster.TASK_RUNLEVEL_HIGHEST
-	task.Definition.Principal.UserID = systemUserName
-
-	// clear up all schedules and put them back
-	task.Definition.Triggers = []taskmaster.Trigger{}
-	createSchedules(&task.Definition, schedules)
-
-	_, err := taskService.UpdateTask(taskName, task.Definition)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func createSchedules(task *taskmaster.Definition, schedules []*calendar.Event) {
-	for _, schedule := range schedules {
-		if once, ok := schedule.AsTime(); ok {
-			// one time only
-			task.AddTrigger(taskmaster.TimeTrigger{
-				RandomDelay: period.Period{},
-				TaskTrigger: taskmaster.TaskTrigger{
-					Enabled:       true,
-					StartBoundary: once,
-				},
-			})
-			continue
-		}
-		if schedule.IsDaily() {
-			// recurring daily
-			createDailyTrigger(task, schedule)
-			continue
-		}
-		if schedule.IsWeekly() {
-			createWeeklyTrigger(task, schedule)
-			continue
-		}
-		if schedule.IsMonthly() {
-			createMonthlyTrigger(task, schedule)
-			continue
-		}
-		clog.Warningf("cannot convert schedule '%s' into a task scheduler equivalent", schedule.String())
-	}
-}
-
-func createDailyTrigger(task *taskmaster.Definition, schedule *calendar.Event) {
-	start := schedule.Next(time.Now())
-	// get all recurrences in the same day
-	recurrences := schedule.GetAllInBetween(start, start.Add(24*time.Hour))
-	if len(recurrences) == 0 {
-		clog.Warningf("cannot convert schedule '%s' into a daily trigger", schedule.String())
-		return
-	}
-	// Is it only once a day?
-	if len(recurrences) == 1 {
-		task.AddTrigger(taskmaster.DailyTrigger{
-			DayInterval: 1,
-			TaskTrigger: taskmaster.TaskTrigger{
-				Enabled:       true,
-				StartBoundary: recurrences[0],
-			},
-		})
-		return
-	}
-	// now calculate the difference in between each, and check if they're all the same
-	_, compactDifferences := compileDifferences(recurrences)
-
-	if len(compactDifferences) == 1 {
-		// case with regular repetition
-		interval, _ := period.NewOf(compactDifferences[0])
-		task.AddTrigger(taskmaster.DailyTrigger{
-			DayInterval: 1,
-			TaskTrigger: taskmaster.TaskTrigger{
-				Enabled:       true,
-				StartBoundary: start,
-				RepetitionPattern: taskmaster.RepetitionPattern{
-					RepetitionDuration: getRepetionDuration(start, recurrences),
-					RepetitionInterval: interval,
-				},
-			},
-		})
-		return
-	}
-
-	if len(recurrences) > maxTriggers {
-		clog.Warningf("this task would need more than %d triggers (%d in total), please rethink your triggers definition", maxTriggers, len(recurrences))
-		return
-	}
-	// install them all
-	for _, recurrence := range recurrences {
-		task.AddTrigger(taskmaster.DailyTrigger{
-			DayInterval: 1,
-			TaskTrigger: taskmaster.TaskTrigger{
-				Enabled:       true,
-				StartBoundary: recurrence,
-			},
-		})
-	}
-}
-
-func createWeeklyTrigger(task *taskmaster.Definition, schedule *calendar.Event) {
-	start := schedule.Next(time.Now())
-	// get all recurrences in the same day
-	recurrences := schedule.GetAllInBetween(start, start.Add(24*time.Hour))
-	if len(recurrences) == 0 {
-		clog.Warningf("cannot convert schedule '%s' into a weekly trigger", schedule.String())
-		return
-	}
-	// Is it only once per 24h?
-	if len(recurrences) == 1 {
-		task.AddTrigger(taskmaster.WeeklyTrigger{
-			DaysOfWeek:   taskmaster.DayOfWeek(convertWeekdaysToBitmap(schedule.WeekDay.GetRangeValues())),
-			WeekInterval: 1,
-			TaskTrigger: taskmaster.TaskTrigger{
-				Enabled:       true,
-				StartBoundary: recurrences[0],
-			},
-		})
-		return
-	}
-	// now calculate the difference in between each, and check if they're all the same
-	_, compactDifferences := compileDifferences(recurrences)
-
-	if len(compactDifferences) == 1 {
-		// case with regular repetition
-		interval, _ := period.NewOf(compactDifferences[0])
-		task.AddTrigger(taskmaster.WeeklyTrigger{
-			DaysOfWeek:   taskmaster.DayOfWeek(convertWeekdaysToBitmap(schedule.WeekDay.GetRangeValues())),
-			WeekInterval: 1,
-			TaskTrigger: taskmaster.TaskTrigger{
-				Enabled:       true,
-				StartBoundary: start,
-				RepetitionPattern: taskmaster.RepetitionPattern{
-					RepetitionDuration: getRepetionDuration(start, recurrences),
-					RepetitionInterval: interval,
-				},
-			},
-		})
-		return
-	}
-
-	if len(recurrences) > maxTriggers {
-		clog.Warningf("this task would need more than %d triggers (%d in total), please rethink your triggers definition", maxTriggers, len(recurrences))
-		return
-	}
-	// install them all
-	for _, recurrence := range recurrences {
-		task.AddTrigger(taskmaster.WeeklyTrigger{
-			DaysOfWeek:   taskmaster.DayOfWeek(convertWeekdaysToBitmap(schedule.WeekDay.GetRangeValues())),
-			WeekInterval: 1,
-			TaskTrigger: taskmaster.TaskTrigger{
-				Enabled:       true,
-				StartBoundary: recurrence,
-			},
-		})
-	}
-}
-
-func createMonthlyTrigger(task *taskmaster.Definition, schedule *calendar.Event) {
-	start := schedule.Next(time.Now())
-	// get all recurrences in the same day
-	recurrences := schedule.GetAllInBetween(start, start.Add(24*time.Hour))
-	if len(recurrences) == 0 {
-		clog.Warningf("cannot convert schedule '%s' into a monthly trigger", schedule.String())
-		return
-	}
-
-	if len(recurrences) > maxTriggers {
-		clog.Warningf("this task would need more than %d triggers (%d in total), please rethink your triggers definition", maxTriggers, len(recurrences))
-		return
-	}
-	// install them all
-	for _, recurrence := range recurrences {
-		if schedule.WeekDay.HasValue() && schedule.Day.HasValue() {
-			clog.Warningf("task scheduler does not support a day of the month and a day of the week in the same trigger: %s", schedule.String())
-			return
-		}
-		if schedule.WeekDay.HasValue() {
-			task.AddTrigger(taskmaster.MonthlyDOWTrigger{
-				DaysOfWeek:   taskmaster.DayOfWeek(convertWeekdaysToBitmap(schedule.WeekDay.GetRangeValues())),
-				WeeksOfMonth: taskmaster.AllWeeks,
-				MonthsOfYear: taskmaster.Month(convertMonthsToBitmap(schedule.Month.GetRangeValues())),
-				TaskTrigger: taskmaster.TaskTrigger{
-					Enabled:       true,
-					StartBoundary: recurrence,
-				},
-			})
-			continue
-		}
-		task.AddTrigger(taskmaster.MonthlyTrigger{
-			DaysOfMonth:  taskmaster.DayOfMonth(convertDaysToBitmap(schedule.Day.GetRangeValues())),
-			MonthsOfYear: taskmaster.Month(convertMonthsToBitmap(schedule.Month.GetRangeValues())),
-			TaskTrigger: taskmaster.TaskTrigger{
-				Enabled:       true,
-				StartBoundary: recurrence,
-			},
-		})
-	}
 }
 
 // Delete a task
 func Delete(title, subtitle string) error {
-	if !IsConnected() {
-		return ErrNotConnected
-	}
-
 	taskName := getTaskPath(title, subtitle)
-	err := taskService.DeleteTask(taskName)
-	if err != nil {
-		if strings.Contains(err.Error(), "doesn't exist") || strings.Contains(err.Error(), "cannot find") {
-			return fmt.Errorf("%w: %s", ErrNotRegistered, taskName)
-		}
-		return err
-	}
-	return nil
+	_, err := deleteTask(taskName)
+	return err
 }
 
 // Status returns the status of a task
 func Status(title, subtitle string) error {
-	if !IsConnected() {
-		return ErrNotConnected
-	}
-
 	taskName := getTaskPath(title, subtitle)
-	registeredTask, err := taskService.GetRegisteredTask(taskName)
+	info, err := getTaskInfo(taskName)
 	if err != nil {
-		// if there's an error here, it is very likely that the task is not registered
-		return fmt.Errorf("%s: %w: %s", taskName, ErrNotRegistered, err)
+		return err
+	}
+	if len(info) < 2 {
+		return ErrNotRegistered
 	}
 	writer := tabwriter.NewWriter(term.GetOutput(), 2, 2, 2, ' ', tabwriter.AlignRight)
-	fmt.Fprintf(writer, "Task:\t %s\n", registeredTask.Path)
-	fmt.Fprintf(writer, "User:\t %s\n", registeredTask.Definition.Principal.UserID)
-	if len(registeredTask.Definition.Actions) > 0 {
-		if action, ok := registeredTask.Definition.Actions[0].(taskmaster.ExecAction); ok {
-			fmt.Fprintf(writer, "Working Dir:\t %v\n", action.WorkingDir)
-			fmt.Fprintf(writer, "Exec:\t %v\n", action.Path+" "+action.Args)
-		}
-	}
-	fmt.Fprintf(writer, "Enabled:\t %v\n", registeredTask.Enabled)
-	fmt.Fprintf(writer, "State:\t %s\n", registeredTask.State.String())
-	fmt.Fprintf(writer, "Missed runs:\t %d\n", registeredTask.MissedRuns)
-	fmt.Fprintf(writer, "Last Run Time:\t %v\n", registeredTask.LastRunTime)
-	fmt.Fprintf(writer, "Last Result:\t %d\n", registeredTask.LastTaskResult)
-	fmt.Fprintf(writer, "Next Run Time:\t %v\n", registeredTask.NextRunTime)
+	fmt.Fprintf(writer, "Task:\t %s\n", getFirstField(info, "TaskName"))
+	fmt.Fprintf(writer, "User:\t %s\n", getFirstField(info, "Run As User"))
+	fmt.Fprintf(writer, "Working Dir:\t %v\n", getFirstField(info, "Start In"))
+	fmt.Fprintf(writer, "Command:\t %v\n", getFirstField(info, "Task To Run"))
+	fmt.Fprintf(writer, "Status:\t %s\n", getFirstField(info, "Status"))
+	fmt.Fprintf(writer, "Last Run Time:\t %v\n", getFirstField(info, "Last Run Time"))
+	fmt.Fprintf(writer, "Last Result:\t %s\n", getFirstField(info, "Last Result"))
+	fmt.Fprintf(writer, "Next Run Time:\t %v\n", getFirstField(info, "Next Run Time"))
 	writer.Flush()
 	return nil
 }
 
 func Registered() ([]Config, error) {
-	if !IsConnected() {
-		return nil, ErrNotConnected
-	}
-
-	tasks, err := taskService.GetRegisteredTasks()
+	raw, err := listRegisteredTasks()
 	if err != nil {
 		return nil, err
 	}
-	configs := make([]Config, 0, len(tasks))
-	for _, task := range tasks {
-		if !strings.HasPrefix(task.Path, tasksPath) {
+	all, err := getCSV(bytes.NewBuffer(raw))
+	if err != nil {
+		return nil, err
+	}
+	list := make([]string, 0)
+	for _, taskLine := range all {
+		if strings.HasPrefix(taskLine[0], tasksPath) {
+			list = append(list, taskLine[0])
+		}
+	}
+	slices.Sort(list)
+	list = slices.Compact(list)
+
+	configs := make([]Config, 0, len(list))
+	for _, taskPath := range list {
+		info, err := getTaskInfo(taskPath)
+		if err != nil {
+			clog.Errorf("loading task %q: %w", taskPath, err)
 			continue
 		}
-		taskPath := strings.TrimPrefix(task.Path, tasksPath)
-		parts := strings.Split(taskPath, " ")
+		if len(info) < 2 {
+			continue
+		}
+		taskName := strings.TrimPrefix(taskPath, tasksPath)
+		parts := strings.Split(taskName, " ")
 		if len(parts) < 2 {
-			clog.Warningf("cannot parse task path: %s", task.Path)
+			clog.Warningf("cannot parse task path: %s", taskPath)
 			continue
 		}
 		profileName := strings.Join(parts[:len(parts)-1], " ")
 		commandName := parts[len(parts)-1]
-		cfg := Config{
-			ProfileName:    profileName,
-			CommandName:    commandName,
-			JobDescription: task.Definition.RegistrationInfo.Description,
+		commandLine := getFirstField(info, "Task To Run")
+		command, args, _ := strings.Cut(commandLine, " ")
+		config := Config{
+			ProfileName:      profileName,
+			CommandName:      commandName,
+			JobDescription:   getFirstField(info, "Comment"),
+			WorkingDirectory: getFirstField(info, "Start In"),
+			Command:          command,
+			Arguments:        args,
 		}
-		if len(task.Definition.Actions) > 0 {
-			if action, ok := task.Definition.Actions[0].(taskmaster.ExecAction); ok {
-				cfg.WorkingDirectory = action.WorkingDir
-				cfg.Command = action.Path
-				cfg.Arguments = action.Args
-			}
-		}
-		configs = append(configs, cfg)
+		fmt.Printf("%+v\n", config)
+		configs = append(configs, config)
 	}
 	return configs, nil
 }
 
-func convertWeekdaysToBitmap(weekdays []int) uint16 {
-	if len(weekdays) == 0 {
-		return 0
-	}
-	var bitmap uint16
-	for _, weekday := range weekdays {
-		bitmap |= getWeekdayBit(weekday)
-	}
-	return bitmap
-}
-
-func getWeekdayBit(weekday int) uint16 {
-	switch weekday {
-	case 0:
-		return 1
-	case 1:
-		return 2
-	case 2:
-		return 4
-	case 3:
-		return 8
-	case 4:
-		return 16
-	case 5:
-		return 32
-	case 6:
-		return 64
-	case 7:
-		// Sunday is the first day of the week
-		return 1
-	}
-	return 0
-}
-
-func convertMonthsToBitmap(months []int) uint16 {
-	if months == nil {
-		return 0
-	}
-	if len(months) == 0 {
-		// all values
-		return (1 << 12) - 1
-	}
-	var bitmap uint16
-	for _, month := range months {
-		if month > 0 && month <= 12 {
-			bitmap |= 1 << (month - 1)
-		}
-	}
-	return bitmap
-}
-
-func convertDaysToBitmap(days []int) uint32 {
-	if days == nil {
-		return 0
-	}
-	if len(days) == 0 {
-		// every day
-		return (1 << 31) - 1
-	}
-	var bitmap uint32
-	for _, day := range days {
-		if day > 0 && day <= 31 {
-			bitmap |= 1 << (day - 1)
-		}
-	}
-	return bitmap
-}
-
 func getTaskPath(profileName, commandName string) string {
 	return fmt.Sprintf("%s%s %s", tasksPath, profileName, commandName)
+}
+
+func getFirstField(data [][]string, fieldName string) string {
+	index := slices.Index(data[0], fieldName)
+	if index < 0 {
+		return ""
+	}
+	return data[1][index]
 }
